Item brand CRUD completed

+++ src/main/java/com/vnc/officeManagementApp/Repository/ItemRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.vnc.officeManagementApp.Models.Items;
 
-public interface ItemRepository extends JpaRepository<Items, Long> {
+public interface ItemRepository extends JpaRepository<Items, Integer> {
 
 }
