+++ src/main/java/com/vnc/officeManagementApp/Repository/RolesRepository.java
@@ -3,5 +3,5 @@ package com.vnc.officeManagementApp.Repository;
 import com.vnc.officeManagementApp.Models.Roles;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolesRepository extends JpaRepository<Roles, Long> {
+public interface RolesRepository extends JpaRepository<Roles, Integer> {
 }
